workspace/py-workspace/armon/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in requirements.testing.in
#
attrs==20.2.0             # via -r requirements.in, pytest
iniconfig==1.0.1          # via pytest
packaging==20.4           # via pytest
pluggy==0.13.1            # via pytest
py==1.9.0                 # via pytest
pyparsing==2.4.7          # via packaging
pytest==6.1.0             # via -r requirements.testing.in
redis==3.5.3              # via -r requirements.in
six==1.15.0               # via packaging
toml==0.10.1              # via pytest
